UIQM-415 Make update request when creating bib to save links
